Further UI development
git-svn-id: svn://localhost/EGE@683 1e37b749-e30f-da40-b4df-3e99ea2a9c57

+++ trunk/Sources/Win32/TypesWin32.h
@@ -3,8 +3,8 @@
 
 //--------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-typedef signed long         s32;            // i
-typedef unsigned long       u32;            // ui
+typedef int                 s32;            // i
+typedef unsigned int        u32;            // ui
 typedef short               s16;            // si
 typedef unsigned short      u16;            // su
 typedef char                s8;             // c
